requirements_unified.txt
# Core Web Framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
aiofiles==23.2.1
jinja2==3.1.2

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# Data Models & Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Document Parsing
python-docx==1.1.0
PyPDF2==3.0.1
PyMuPDF==1.23.8
pdfplumber==0.10.3

# Data Processing
pandas==2.1.4
numpy==1.24.3
python-dateutil==2.8.2

# NLP & ML Core
scikit-learn==1.3.2
nltk==3.8.1
spacy==3.7.2
textstat==0.7.3

# Deep Learning & Transformers
torch==2.1.2
transformers==4.36.2
sentence-transformers==2.2.2

# Security & Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Additional NLP Models (download separately)
# python -m spacy download en_core_web_sm
# python -m nltk.downloader punkt stopwords wordnet averaged_perceptron_tagger
